move DumyPar to a shared location
